requirements.txt
absl-py==0.9.0
alembic==1.4.2
apache-airflow==1.10.12
apispec==1.3.3
app-version==1.0.1
appnope==0.1.0
argcomplete==1.11.1
astroid==2.3.3
astunparse==1.6.3
attrs==19.3.0
Babel==2.8.0
backcall==0.1.0
bleach==3.1.0
blinker==1.4
bokeh==1.4.0
brotlipy==0.7.0
bson==0.5.10
cached-property==1.5.1
cachetools==4.0.0
cattrs==1.0.0
certifi==2020.6.20
cffi @ file:///opt/concourse/worker/volumes/live/b9607b09-b777-4ff7-53dc-287727eb8574/volume/cffi_1600699191154/work
chardet==3.0.4
Click==7.0
cloudpickle==1.2.2
colorama==0.4.3
colorlog==4.0.2
conda==4.9.2
conda-package-handling @ file:///opt/concourse/worker/volumes/live/a7e34989-4c54-4cb6-4156-4e58ee270730/volume/conda-package-handling_1603018121300/work
configparser==3.5.3
croniter==0.3.31
cryptography @ file:///opt/concourse/worker/volumes/live/aeb63a26-659e-4edb-5405-74ba8e0c76f2/volume/cryptography_1601046839724/work
cycler==0.10.0
Cython==0.29.14
dash==1.9.1
dash-core-components==1.8.1
dash-html-components==1.0.2
dash-renderer==1.2.4
dash-table==4.6.1
dask==2.13.0
decorator==4.4.1
defusedxml==0.6.0
dill==0.3.1.1
distributed==2.30.0
dnspython==1.16.0
docutils==0.16
dominate==2.5.1
email-validator==1.1.1
entrypoints==0.3
Flask==1.1.1
Flask-Admin==1.5.4
Flask-AppBuilder==2.3.4
Flask-Babel==1.0.0
Flask-Bootstrap==3.3.7.1
Flask-Caching==1.3.3
Flask-Compress==1.4.0
Flask-JWT-Extended==3.24.1
Flask-Login==0.4.1
Flask-Mail==0.9.1
Flask-OpenID==1.2.5
Flask-SQLAlchemy==2.4.1
flask-swagger==0.2.13
Flask-WTF==0.14.3
fsspec==0.6.2
funcsigs==1.0.2
future==0.18.2
gast==0.3.3
gcsfs==0.6.0
get==2019.4.13
google-api-core==1.23.0
google-api-python-client==1.8.3
google-auth==1.23.0
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.4.1
google-cloud==0.34.0
google-cloud-bigquery==1.23.1
google-cloud-bigquery-storage==1.1.0
google-cloud-bigtable==1.5.1
google-cloud-container==2.1.0
google-cloud-core==1.4.3
google-cloud-dlp==2.0.0
google-cloud-language==2.0.0
google-cloud-logging==1.15.0
google-cloud-secret-manager==1.0.0
google-cloud-spanner==2.0.0
google-cloud-speech==2.0.0
google-cloud-storage==1.33.0
google-cloud-texttospeech==1.0.1
google-cloud-translate==3.0.1
google-cloud-videointelligence==1.16.0
google-cloud-vision==2.0.0
google-crc32c==1.0.0
google-oauth==1.0.1
google-pasta==0.2.0
google-resumable-media==1.1.0
googleapis-common-protos==1.51.0
graphframes==0.6
graphviz==0.14
grpc-google-iam-v1==0.12.3
grpcio==1.28.1
grpcio-gcp==0.2.2
gunicorn==19.10.0
h5py==2.10.0
HeapDict==1.0.1
httplib2==0.18.1
idna @ file:///tmp/build/80754af9/idna_1593446292537/work
ipykernel==5.1.4
ipython==7.12.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
iso8601==0.1.12
isort==4.3.21
itsdangerous==1.1.0
jedi==0.16.0
Jinja2==2.10.3
joblib==0.14.1
json-merge-patch==0.2
json5==0.9.3
jsonify==0.5
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==5.3.4
jupyter-console==6.1.0
jupyter-core==4.6.1
jupyterlab==2.0.1
jupyterlab-server==1.0.7
kiwisolver==1.1.0
lazy-object-proxy==1.4.3
libcst==0.3.13
locket==0.2.0
lockfile==0.12.2
lxml==4.5.2
Mako==1.1.2
Markdown==2.6.11
MarkupSafe==1.1.1
marshmallow==2.21.0
marshmallow-enum==1.5.1
marshmallow-sqlalchemy==0.23.0
matplotlib==3.2.1
mccabe==0.6.1
mistune==0.8.4
msgpack==1.0.0
mypy-extensions==0.4.3
nbconvert==5.6.1
nbformat==5.0.4
nose==1.3.7
notebook==6.0.3
ntlm-auth==1.5.0
numpy==1.18.2
oauthlib==3.1.0
opt-einsum==3.2.1
packaging==20.1
pandas==1.0.3
pandas-gbq==0.14.1
pandocfilters==1.4.2
parso==0.6.0
partd==1.1.0
pendulum==1.4.4
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.0.0
plotly==4.5.0
post==2019.4.13
prison==0.1.3
prometheus-client==0.7.1
prompt-toolkit==3.0.3
proto-plus==1.11.0
protobuf==3.14.0
psutil==5.7.0
ptyprocess==0.6.0
public==2019.4.13
py4j==0.10.9
pyarrow==2.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pybigquery==0.4.15
pycosat==0.6.3
pycparser @ file:///tmp/build/80754af9/pycparser_1594388511720/work
pydata-google-auth==1.1.0
Pygments==2.5.2
PyJWT==1.7.1
pylint==2.4.4
pymongo==3.11.0
pyodbc==4.0.28
pyOpenSSL @ file:///tmp/build/80754af9/pyopenssl_1594392929924/work
pyparsing==2.4.6
pyrsistent==0.15.7
PySocks @ file:///opt/concourse/worker/volumes/live/85a5b906-0e08-41d9-6f59-084cee4e9492/volume/pysocks_1594394636991/work
pyspark==3.0.0
python-daemon==2.1.2
python-dateutil==2.8.1
python-editor==1.0.4
python-gnupg==0.4.6
python-nvd3==0.15.0
python-slugify==4.0.1
python3-openid==3.1.0
pytz==2019.3
pytzdata==2019.3
PyYAML==5.3
pyzmq==18.1.1
qtconsole==4.6.0
query-string==2019.4.13
request==2019.4.13
requests @ file:///tmp/build/80754af9/requests_1592841827918/work
requests-ntlm==1.1.0
requests-oauthlib==1.3.0
requests-toolbelt==0.9.1
retrying==1.3.3
rsa==4.0
ruamel-yaml==0.15.87
scikit-learn==0.22.1
scipy==1.4.1
seaborn==0.10.0
Send2Trash==1.5.0
setproctitle==1.1.10
SharePlum==0.5.1
six @ file:///opt/concourse/worker/volumes/live/5b31cb27-1e37-4ca5-6e9f-86246eb206d2/volume/six_1605205320872/work
sortedcontainers==2.2.2
SQLAlchemy==1.3.13
SQLAlchemy-JSONField==0.9.0
SQLAlchemy-Utils==0.36.6
style==1.1.0
tabulate==0.8.7
tblib==1.7.0
tenacity==4.12.0
termcolor==1.1.0
terminado==0.8.3
testpath==0.4.4
text-unidecode==1.3
thrift==0.13.0
toolz==0.11.1
tornado==6.0.3
tqdm @ file:///tmp/build/80754af9/tqdm_1602185206534/work
traitlets==4.3.3
typing-extensions==3.7.4.3
typing-inspect==0.6.0
tzlocal==1.5.1
unicodecsv==0.14.1
update==0.0.1
uritemplate==3.0.1
urllib3 @ file:///tmp/build/80754af9/urllib3_1603305693037/work
validate-email==1.3
visitor==0.1.3
wcwidth==0.1.8
webencodings==0.5.1
Werkzeug==0.16.1
widgetsnbextension==3.5.1
wrapt==1.11.2
WTForms==2.2.1
XlsxWriter==1.2.7
zict==2.0.0
zope.deprecation==4.4.0
